services/document-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.27.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.6
langchain==0.3.20
langchain-google-genai==2.1.11
tiktoken==0.5.2
pypdf==4.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10
sentence-transformers==2.3.1
beautifulsoup4==4.12.3
httpx==0.26.0
redis==5.0.1
prometheus-fastapi-instrumentator==7.0.0
boto3==1.34.162
pdf2image==1.17.0
